feat(calculator): Add calculator tool
This commit adds a new calculator tool that can be used to evaluate mathematical expressions. The
tool is implemented as a shell script that uses `bc` to perform the calculations.

The tool is added to the list of available tools and the function definition is added to the
functions.json file.